Move Traversable1 into semantic-ast.
