+++ semantic-java/semantic-java.cabal
@@ -27,6 +27,7 @@ library
   build-depends:
       base              >= 4.13 && < 5
     , fused-effects    ^>= 1.0
+    , semantic-ast
     , semantic-codegen
     , semantic-source  ^>= 0.0.2
     , semantic-tags    ^>= 0.0
